+++ libUDB/events.h
@@ -29,7 +29,7 @@
 #define INVALID_HANDLE 0xFFFF
 
 void init_events(void);
-void trigger_event(unsigned int hEvent);
+void trigger_event(uint16_t hEvent);
 
 typedef enum eventP
 {
@@ -42,12 +42,12 @@ typedef struct tagEVENT
 {
 	boolean eventPending;
 	void (*event_callback) (void);
-	int priority;
+	int16_t priority;
 } EVENT;
 
-unsigned int register_event( void (*event_callback) (void) );
+uint16_t register_event( void (*event_callback) (void) );
 
-unsigned int register_event_p(void (*event_callback) (void), eventPriority priority );
+uint16_t register_event_p(void (*event_callback) (void), eventPriority priority );
 
 
 #endif
