+++ libUDB/mpu6000.c
@@ -30,13 +30,13 @@
 #include "../libDCM/libDCM_internal.h"
 
 //Sensor variables
-unsigned int mpu_data[8], mpuCnt = 0;
+uint16_t mpu_data[8], mpuCnt = 0;
 bool mpuDAV = false;
 
 struct ADchannel udb_xaccel, udb_yaccel, udb_zaccel; // x, y, and z accelerometer channels
 struct ADchannel udb_xrate, udb_yrate, udb_zrate; // x, y, and z gyro channels
 struct ADchannel mpu_temp;
-int vref_adj;
+int16_t vref_adj;
 
 // MPU6000 Initialization and configuration
